Added test cases for the tree control issue (#69). Fixed spelling in DeploymentConfiguration.
